Moved vue augmentation into separate file
