+++ src/PrismicVue.ts
@@ -1,5 +1,4 @@
 import * as prismicJS from 'prismic-javascript'
-import {DefaultClient} from 'prismic-javascript/d.ts/client'
 import _Vue from 'vue'
 import {PrismicEditButton} from './components/EditButton'
 import {PrismicEmbed} from './components/Embed'
@@ -7,6 +6,7 @@ import {PrismicImage} from './components/Image'
 import {PrismicLink} from './components/Link'
 import {PrismicRichText} from './components/RichText'
 import {IPrismicVueOptions} from './IPrismicVueOptions'
+import './PrismicVueAugment'
 
 export function PrismicVue(Vue: typeof _Vue, options: IPrismicVueOptions) {
   // Register prototypes
@@ -22,19 +22,4 @@ export function PrismicVue(Vue: typeof _Vue, options: IPrismicVueOptions) {
   Vue.component('PrismicImage', PrismicImage)
   Vue.component('PrismicLink', PrismicLink)
   Vue.component('PrismicRichText', PrismicRichText)
-}
-
-// Augment Vue
-declare module 'vue/types/vue' {
-  // tslint:disable-next-line:interface-name
-  interface Vue {
-    $prismic: {
-      Predicates: typeof prismicJS.Predicates,
-      Experiments: typeof prismicJS.Experiments
-      Api: typeof prismicJS.Api
-      client: DefaultClient,
-      getApi: typeof prismicJS.getApi
-      api: typeof prismicJS.api
-    }
-  }
 }
